oublié ce que j'ai fait

+++ src/Controller/RoomController.php
@@ -3,8 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Room;
+use App\Entity\Commentaire;
 use App\Form\RoomType;
 use App\Repository\RoomRepository;
+use App\Repository\CommentaireRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -65,18 +67,64 @@ class RoomController extends AbstractController
     /**
      * @Route("/{id}", name="room_show", methods={"GET"})
      */
-    public function show(Room $room): Response
+    public function show(Room $room, CommentaireRepository $commentaireRepository): Response
     {
+        if (!$this->getUser()->getOwner()->owns($room) && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $this->get('session')->getFlashBag()->add('error', 'Vous n\'avez pas le droit d\'accéder à ceci');
+            return $this->redirectToRoute('room_index');
+        }
+
+        $commentaires = $commentaireRepository->findBy(['status' => 'EN ATTENTE DE CONFIRMATION', 'room' => $room]);
         return $this->render('room/show.html.twig', [
             'room' => $room,
+            'commentaires' => $commentaires,
+            'owner' => $room->getOwner()
         ]);
     }
 
+    /**
+     * @Route("/commentaire/accept/{id}", name="room_commentaire_accept")
+     */
+    public function commentaireAccept(Commentaire $commentaire): Response
+    {
+        if (!$this->getUser()->getOwner()->owns($commentaire->getRoom())) {
+            $this->get('session')->getFlashBag()->add('error', 'Vous n\'avez pas le droit de faire ceci');
+            return $this->redirectToRoute('room_index');
+        }
+
+        $commentaire->setStatus('VALIDE');
+        $this->getDoctrine()->getManager()->flush();
+
+        $this->get('session')->getFlashBag()->add('success', 'Commentaire validé.');
+        return $this->redirectToRoute('room_show', ['id' => $commentaire->getRoom()->getId()]);
+    }
+
+    /**
+     * @Route("/commentaire/report/{id}", name="room_commentaire_report")
+     */
+    public function commentaireReport(Commentaire $commentaire): Response
+    {
+        if (!$this->getUser()->getOwner()->owns($commentaire->getRoom())) {
+            $this->get('session')->getFlashBag()->add('error', 'Vous n\'avez pas le droit de faire ceci');
+            return $this->redirectToRoute('room_index');
+        }
+
+        $commentaire->setStatus('EN ATTENTE DE MODERATION');
+        $this->getDoctrine()->getManager()->flush();
+
+        $this->get('session')->getFlashBag()->add('message', 'Le commentaire a bien été signalé et sera traité sous peu.');
+        return $this->redirectToRoute('room_show', ['id' => $commentaire->getRoom()->getId()]);
+    }
     /**
      * @Route("/{id}/edit", name="room_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Room $room): Response
     {
+        if (!$this->getUser()->getOwner()->owns($room) && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $this->get('session')->getFlashBag()->add('error', 'Vous n\'avez pas le droit d\'accéder à ceci');
+            return $this->redirectToRoute('room_index');
+        }
+
         $form = $this->createForm(RoomType::class, $room);
         $form->handleRequest($request);
 
@@ -97,6 +145,11 @@ class RoomController extends AbstractController
      */
     public function delete(Request $request, Room $room): Response
     {
+        if (!$this->getUser()->getOwner()->owns($room) && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            $this->get('session')->getFlashBag()->add('error', 'Vous n\'avez pas le droit d\'accéder à ceci');
+            return $this->redirectToRoute('room_index');
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$room->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($room);
